+++ lib/pragma/rails/controller.rb
@@ -23,7 +23,8 @@ module Pragma
           result = operation_const.call(
             operation_params,
             'current_user' => operation_user,
-            'policy.context' => policy_context
+            'policy.context' => policy_context,
+            'rack.request' => Rack::Request.new(request.env),
           )
 
           fail NoResponseError unless result['result.response']