+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- Started raising a `NoResponseError` if an operation doesn't return a `result.response` skill 
+- `NoResponseError` is now raised if an operation doesn't return a `result.response` skill 
 - `rack.request` is now passed to operations
 
 ## [2.2.0]