+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.3.0]
+
 ### Added
 
 - `NoResponseError` is now raised if an operation doesn't return a `result.response` skill 
@@ -46,7 +48,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 First Pragma 2 release.
 
-[Unreleased]: https://github.com/pragmarb/pragma-rails/compare/v2.2.0...HEAD
+[Unreleased]: https://github.com/pragmarb/pragma-rails/compare/v2.3.0...HEAD
+[2.3.0]: https://github.com/pragmarb/pragma-rails/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/pragmarb/pragma-rails/compare/v2.1.1...v2.2.0
 [2.1.1]: https://github.com/pragmarb/pragma-rails/compare/v2.1.0...v2.1.1
 [2.1.0]: https://github.com/pragmarb/pragma-rails/compare/v2.0.0...v2.1.0